remove redundant file
