+++ src/preprocessing/index_builder.py
@@ -1,9 +1,9 @@
 import argparse
+import json
 import os
 from pathlib import Path
 
 from dotenv import load_dotenv
-from filter_reactant_fg import filter_reactant_fg
 from llama_index import ServiceContext, VectorStoreIndex
 from llama_index.embeddings import AzureOpenAIEmbedding
 from llama_index.llms import AzureOpenAI
@@ -55,6 +55,39 @@ def build_index(patents_path, save_path, prefix="", chunksize=1000, overlap=20):
     index = VectorStoreIndex(nodes, service_context=service_context, show_progress=True)
     # By default saves to save_path
     index.storage_context.persist(persist_dir=save_path)
+
+
+def filter_reactant_fg(json_file):
+    fg_list = [
+        "alpha aryl carboxylic acid",
+        "acidic groups",
+        "primary amine",
+        "secondary amine",
+        "aldehyde",
+        "ketone",
+        "alcohol",
+    ]
+    with open(json_file, "r") as f:
+        json_content = json.load(f)
+    reaction_list = json_content["reactionList"].get("reaction")
+    good_reactions = []
+    if reaction_list:
+        for reaction in reaction_list:
+            reactant_fg = []
+            if "reactant" in reaction["reactantList"].keys():
+                reactant_list = reaction["reactantList"].get("reactant")
+                # among reactant we want at list one of the fg_list exists
+                if isinstance(reactant_list, list):
+                    for reactant in reactant_list:
+                        if "funcgroups" in reactant.keys():
+                            reactant_fg.extend(reactant["funcgroups"])
+                elif isinstance(reactant_list, dict):
+                    reactant_fg.extend(reactant_list["funcgroups"])
+            checker = any([fg in reactant_fg for fg in fg_list])
+            if checker:
+                good_reactions.append(reaction)
+    print(f"after removing: {len(good_reactions)}")
+    return good_reactions
 
 
 if __name__ == "__main__":
